Extract column schema into a CatalogSchema class
